+++ src/http/oauth1.ts
@@ -204,6 +204,7 @@ export function createOAuth1HttpClient(
    */
   return Object.defineProperty(
     {
+      normalizeCustomData: options.normalizeCustomData,
       ...httpWithAuth,
       extraAuthMethods: {
         authenticate,