+++ src/http/oauth2.ts
@@ -228,6 +228,7 @@ export function createOAuth2HttpClient(
    */
   return Object.defineProperty(
     {
+      normalizeCustomData: options.normalizeCustomData,
       ...httpWithAuth,
       extraAuthMethods: {
         authenticate,