+++ src/http/utils.ts
@@ -5,29 +5,39 @@ export function mapObjIndexed(fn, object): Record<string, unknown> {
   );
 }
 
-export const recursiveMap =
-  (fn, skipData = false) =>
-  obj => {
-    // needed for arrays with strings/numbers etc
-    if (obj === null || typeof obj !== 'object') {
-      return obj;
+// WARNING: skipData is old behaviour that we can not remove yet for backwards compatibility
+// Removing it would affect raw calls to the data-service
+export const recursiveMap = (fn, obj, ignoreKeys = [], skipData = false) => {
+  // needed for arrays with strings/numbers etc
+  if (obj === null || typeof obj !== 'object' || ignoreKeys.includes('*')) {
+    return obj;
+  }
+
+  if (Array.isArray(obj)) {
+    return obj.map(value => recursiveMap(fn, value, ignoreKeys, skipData));
+  }
+
+  return mapObjIndexed((value, key) => {
+    if (typeof value !== 'object' || ignoreKeys.includes(key)) {
+      return fn(value, key);
     }
 
-    return Array.isArray(obj)
-      ? obj.map(recursiveMap(fn, skipData))
-      : mapObjIndexed((value, key) => {
-          if (typeof value !== 'object') {
-            return fn(value, key);
-          }
+    // WARNING: this is old stuff that we could not remove yet for backwards compatibility
+    // if recursively mapping and we find a data property that
+    // is an object, return the object.
+    if (skipData && key === 'data' && !Array.isArray(value)) {
+      return value;
+    }
 
-          // if recursively mapping and we find a data property that
-          // is an object, return the object.
-          if (skipData && key === 'data' && !Array.isArray(value)) {
-            return value;
-          }
-          return recursiveMap(fn, skipData)(value);
-        }, obj);
-  };
+    // Filter all keys in `ignoreKeys` that start with the current `key` before the first `.`
+    // Remove the current `key` from those keys before going into the next iteration.
+    const partialIgnoreKeys = ignoreKeys
+      .filter(k => k.startsWith(`${key}.`))
+      .map(k => k.split('.').slice(1).join('.'));
+
+    return recursiveMap(fn, value, partialIgnoreKeys, skipData);
+  }, obj);
+};
 
 /**
  * See if an object (`val`) is an instance of the supplied constructor. This
@@ -36,16 +46,45 @@ export const recursiveMap =
 function is(Ctor, value) {
   return (value != null && value.constructor === Ctor) || value instanceof Ctor;
 }
+/**
+ * Recursively renames keys in an object using a provided function.
+ * @param fn A function that takes a string (key) and returns a string (new key).
+ * @param obj The object to rename they keys in.
+ * @param ignoreKeys An array of keys to ignore during the renaming process.
+ * @returns An object with keys renamed according to the provided function.
+ */
+export function recursiveRenameKeys(
+  fn: { (arg: string): string },
+  obj,
+  ignoreKeys: string[] = []
+) {
+  // If ignoreKeys includes the '*', we don't need to convert anything in this object
+  if (ignoreKeys.includes('*')) return obj;
 
-export function recursiveRenameKeys(fn: { (arg: string): string }, obj) {
+  // If the object is an array, recursively apply the function to each element in the array.
   if (Array.isArray(obj)) {
-    return obj.map(value => recursiveRenameKeys(fn, value));
+    return obj.map(value => recursiveRenameKeys(fn, value, ignoreKeys));
   }
 
   if (is(Object, obj)) {
     return Object.keys(obj).reduce((memo, key) => {
-      if (is(Object, obj[key]) && !is(Date, obj[key])) {
-        return { ...memo, [fn(key)]: recursiveRenameKeys(fn, obj[key]) };
+      // Check that `obj[key]` is an object
+      // If it is date or `ignoreKeys` includes the current key, don't continue the recursion
+      if (
+        is(Object, obj[key]) &&
+        !is(Date, obj[key]) &&
+        !ignoreKeys.includes(key)
+      ) {
+        // Filter all keys in `ignoreKeys` that start with the current `key` before the first `.`
+        // Remove the current `key` from those keys before going into the next iteration.
+        const partialIgnoreKeys = ignoreKeys
+          .filter(k => k.startsWith(`${key}.`))
+          .map(k => k.split('.').slice(1).join('.'));
+
+        return {
+          ...memo,
+          [fn(key)]: recursiveRenameKeys(fn, obj[key], partialIgnoreKeys),
+        };
       }
       return { ...memo, [fn(key)]: obj[key] };
     }, {});
@@ -73,15 +112,17 @@ export function decamelize(string: string): string {
 }
 
 export function camelizeKeys(
-  object: Record<string, unknown>
+  object: Record<string, unknown>,
+  ignoreKeys?: string[]
 ): Record<string, unknown> {
-  return recursiveRenameKeys(camelize, object);
+  return recursiveRenameKeys(camelize, object, ignoreKeys);
 }
 
 export function decamelizeKeys(
-  object: Record<string, unknown>
+  object: Record<string, unknown>,
+  ignoreKeys?: string[]
 ): Record<string, unknown> {
-  return recursiveRenameKeys(decamelize, object);
+  return recursiveRenameKeys(decamelize, object, ignoreKeys);
 }
 
 /**