+++ src/services/http-client.ts
@@ -5,10 +5,14 @@ import {
   HttpRequestConfig,
   HttpResponseError,
 } from '../http/types';
+import { setCustomKeysConfigurationInRequestConfig } from './helpers';
 
 interface HttpClientOptions {
   basePath: string;
-  transformRequestData?(args: Record<string, unknown>): Record<string, unknown>;
+  transformRequestData?(
+    args: Record<string, unknown>,
+    options?: HttpRequestConfig
+  ): Record<string, unknown>;
 }
 
 interface Options {
@@ -37,65 +41,103 @@ const defaultRetryConfig = {
 
 const httpClient = ({
   basePath,
-  transformRequestData = data => data,
+  transformRequestData = (data, _options) => data,
 }: HttpClientOptions) => ({
   get: (
     axios: HttpInstance,
     url: string,
     { shouldRetry, ...config }: HttpRequestConfig & GetOptions = {}
   ) =>
-    axios.get(`${basePath}${url}`, {
-      ...config,
-      ...(shouldRetry !== false ? { retry: defaultRetryConfig } : {}),
-    }),
-  put: (axios: HttpInstance, url: string, data, config?: HttpRequestConfig) =>
-    axios.put(`${basePath}${url}`, transformRequestData(data), config),
+    axios.get(
+      `${basePath}${url}`,
+      setCustomKeysConfigurationInRequestConfig(axios, {
+        ...config,
+        ...(shouldRetry !== false ? { retry: defaultRetryConfig } : {}),
+      })
+    ),
+  put: (axios: HttpInstance, url: string, data, config?: HttpRequestConfig) => {
+    const requestConfig = setCustomKeysConfigurationInRequestConfig(
+      axios,
+      config
+    );
+    return axios.put(
+      `${basePath}${url}`,
+      transformRequestData(data, requestConfig),
+      requestConfig
+    );
+  },
   post: (
     axios: HttpInstance,
     url: string,
     data,
     config?: HttpRequestConfig,
     options?: Options
-  ) =>
-    axios.post(`${basePath}${url}`, transformRequestData(data), {
-      ...config,
-      ...(options?.gzip
-        ? {
-            transformRequest: [
-              // eslint-disable-next-line no-nested-ternary
-              ...(axios.defaults.transformRequest
-                ? Array.isArray(axios.defaults.transformRequest)
-                  ? axios.defaults.transformRequest
-                  : [axios.defaults.transformRequest]
-                : []),
-              (dataInTransform, headers) => {
-                if (typeof dataInTransform === 'string') {
-                  // eslint-disable-next-line no-param-reassign
-                  headers['Content-Encoding'] = 'gzip';
+  ) => {
+    const requestConfig = setCustomKeysConfigurationInRequestConfig(
+      axios,
+      config
+    );
+    return axios.post(
+      `${basePath}${url}`,
+      transformRequestData(data, requestConfig),
+      {
+        ...requestConfig,
+        ...(options?.gzip
+          ? {
+              transformRequest: [
+                // eslint-disable-next-line no-nested-ternary
+                ...(axios.defaults.transformRequest
+                  ? Array.isArray(axios.defaults.transformRequest)
+                    ? axios.defaults.transformRequest
+                    : [axios.defaults.transformRequest]
+                  : []),
+                (dataInTransform, headers) => {
+                  if (typeof dataInTransform === 'string') {
+                    // eslint-disable-next-line no-param-reassign
+                    headers['Content-Encoding'] = 'gzip';
 
-                  // React Native on Android gzips the data implicitly. DO NOT TOUCH!
-                  if (platform.platform === 'android') {
-                    return dataInTransform;
+                    // React Native on Android gzips the data implicitly. DO NOT TOUCH!
+                    if (platform.platform === 'android') {
+                      return dataInTransform;
+                    }
+
+                    // Nodejs uses the http adapter in Axios. Needs a Buffer with the gzip data
+                    if (platform.platform === 'nodejs') {
+                      return Buffer.from(gzipSync(strToU8(dataInTransform)));
+                    }
+
+                    // In Browser and React Native on iOS
+                    return gzipSync(strToU8(dataInTransform));
                   }
-
-                  // Nodejs uses the http adapter in Axios. Needs a Buffer with the gzip data
-                  if (platform.platform === 'nodejs') {
-                    return Buffer.from(gzipSync(strToU8(dataInTransform)));
-                  }
-
-                  // In Browser and React Native on iOS
-                  return gzipSync(strToU8(dataInTransform));
-                }
-                return dataInTransform;
-              },
-            ],
-          }
-        : {}),
-    }),
+                  return dataInTransform;
+                },
+              ],
+            }
+          : {}),
+      }
+    );
+  },
   delete: (axios: HttpInstance, url: string, config?: HttpRequestConfig) =>
-    axios.delete(`${basePath}${url}`, config),
-  patch: (axios: HttpInstance, url: string, data, config?: HttpRequestConfig) =>
-    axios.patch(`${basePath}${url}`, transformRequestData(data), config),
+    axios.delete(
+      `${basePath}${url}`,
+      setCustomKeysConfigurationInRequestConfig(axios, config)
+    ),
+  patch: (
+    axios: HttpInstance,
+    url: string,
+    data,
+    config?: HttpRequestConfig
+  ) => {
+    const requestConfig = setCustomKeysConfigurationInRequestConfig(
+      axios,
+      config
+    );
+    return axios.patch(
+      `${basePath}${url}`,
+      transformRequestData(data, requestConfig),
+      requestConfig
+    );
+  },
   options: (axios: HttpInstance, url: string) =>
     axios.options(`${basePath}${url}`),
   head: (axios: HttpInstance, url: string) => axios.head(`${basePath}${url}`),