+++ src/services/tasks/tasks.ts
@@ -3,15 +3,15 @@ import type { Task, TasksService } from './types';
 import { rqlBuilder } from '../../rql';
 import { HttpClient } from '../http-client';
 import { findAllGeneric, findAllIterator } from '../helpers';
-import functions from './functions';
-import api from './api';
-import logs from './logs';
-import apiRequests from './apiRequests';
-import schedules from './schedules';
 
 export default (client: HttpClient, httpAuth: HttpInstance): TasksService => ({
   async find(options) {
-    return (await client.get(httpAuth, `/${options?.rql || ''}`, options)).data;
+    return (
+      await client.get(httpAuth, `/${options?.rql || ''}`, {
+        ...options,
+        customResponseKeys: ['data.data'],
+      })
+    ).data;
   },
 
   async findAll(this: TasksService, options) {
@@ -34,16 +34,15 @@ export default (client: HttpClient, httpAuth: HttpInstance): TasksService => ({
   },
 
   async create(requestBody, options) {
-    return (await client.post(httpAuth, '/', requestBody, options)).data;
+    return (
+      await client.post(httpAuth, '/', requestBody, {
+        ...options,
+        customKeys: ['data'],
+      })
+    ).data;
   },
 
   async cancel(taskId, options) {
     return (await client.post(httpAuth, `/${taskId}/cancel`, {}, options)).data;
   },
-
-  schedules: schedules(client, httpAuth),
-  functions: functions(client, httpAuth),
-  api: api(client, httpAuth),
-  logs: logs(client, httpAuth),
-  apiRequests: apiRequests(client, httpAuth),
 });