+++ tests/__helpers__/configuration.ts
@@ -1,3 +1,5 @@
+import { randomHexString } from './utils';
+
 export const generalConfig = {
   data: {
     epicFeatureEnabled: true,
@@ -16,9 +18,34 @@ export const generalConfig = {
   },
 };
 
+export const customGeneralConfig = {
+  data: {
+    epic_feature_enabled: true,
+  },
+  userConfiguration: {
+    epic_feature_enabled: true,
+  },
+  groupConfiguration: {
+    epic_feature_enabled: true,
+  },
+  staffConfiguration: {
+    epic_feature_enabled: true,
+  },
+  patientConfiguration: {
+    epic_feature_enabled: true,
+  },
+};
+
 export const generalConfigResponse = {
   ...generalConfig,
   id: 'abcdef123456789abcdef123',
+  updateTimestamp: '2018-04-24T11:57:44.525Z',
+  creationTimestamp: '2018-04-24T11:57:44.525Z',
+};
+
+export const customGeneralConfigResponse = {
+  ...customGeneralConfig,
+  id: randomHexString(24),
   updateTimestamp: '2018-04-24T11:57:44.525Z',
   creationTimestamp: '2018-04-24T11:57:44.525Z',
 };
@@ -35,6 +62,18 @@ export const groupConfig = {
   },
 };
 
+export const customGroupConfig = {
+  data: {
+    epic_feature_enabled: true,
+  },
+  staffConfiguration: {
+    epic_feature_enabled: true,
+  },
+  patientConfiguration: {
+    epic_feature_enabled: true,
+  },
+};
+
 export const groupConfigResponse = {
   ...groupConfig,
   id: 'abcdef123456789abcdef123',
@@ -42,19 +81,17 @@ export const groupConfigResponse = {
   creationTimestamp: '2018-04-24T11:57:44.525Z',
 };
 
-export const groupConfigInput = {
+export const customGroupConfigResponse = {
+  ...customGroupConfig,
+  id: randomHexString(24),
+  updateTimestamp: '2018-04-24T11:57:44.525Z',
+  creationTimestamp: '2018-04-24T11:57:44.525Z',
+};
+
+export const userConfig = {
   data: {
     epicFeatureEnabled: true,
   },
-  staffConfiguration: {
-    epicFeatureEnabled: true,
-  },
-  patientConfiguration: {
-    epicFeatureEnabled: true,
-  },
-};
-
-export const userEnlistmentsConfig = {
   staffConfigurations: [
     {
       groupId: 'abcdef123456789abcdef123',
@@ -73,16 +110,50 @@ export const userEnlistmentsConfig = {
   ],
 };
 
-export const userConfigInput = {
+export const customUserConfig = {
   data: {
-    epicFeatureEnabled: true,
+    epic_feature_enabled: true,
   },
+  staffConfigurations: [
+    {
+      groupId: randomHexString(24),
+      data: {
+        epic_feature_enabled: true,
+      },
+    },
+  ],
+  patientConfigurations: [
+    {
+      groupId: randomHexString(24),
+      data: {
+        epic_feature_enabled: true,
+      },
+    },
+  ],
 };
 
 export const userConfigResponse = {
-  ...userEnlistmentsConfig,
-  ...userConfigInput,
+  ...userConfig,
   id: 'abcdef123456789abcdef123',
   updateTimestamp: '2018-04-24T11:57:44.525Z',
   creationTimestamp: '2018-04-24T11:57:44.525Z',
+};
+
+export const customUserConfigResponse = {
+  ...customUserConfig,
+  id: randomHexString(24),
+  updateTimestamp: '2018-04-24T11:57:44.525Z',
+  creationTimestamp: '2018-04-24T11:57:44.525Z',
+};
+
+export const groupConfigInput = {
+  data: {
+    epicFeatureEnabled: true,
+  },
+  staffConfiguration: {
+    epicFeatureEnabled: true,
+  },
+  patientConfiguration: {
+    epicFeatureEnabled: true,
+  },
 };