+++ tests/__helpers__/task.ts
@@ -28,23 +28,47 @@ export interface OutputType {
 export const directExecutionResponse = {
   id: '757f191a810c19729de860ae',
   status: 'new',
-  statusChangedTimestamp: new Date('2023-07-03T14:14:50.698Z'),
+  statusChangedTimestamp: '2023-07-03T14:14:50.698Z',
   functionName: 'test',
   data: {
     inputOne: 'value',
     inputTwo: 'value',
     inputThree: 'value',
   },
-  startTimestamp: new Date('2023-07-03T14:14:50.698Z'),
+  startTimestamp: '2023-07-03T14:14:50.698Z',
   tags: ['nightly', 'test'],
   priority: 1,
-  creationTimestamp: new Date('2023-07-03T14:14:50.698Z'),
-  updateTimestamp: new Date('2023-07-03T14:14:50.698Z'),
+  creationTimestamp: '2023-07-03T14:14:50.698Z',
+  updateTimestamp: '2023-07-03T14:14:50.698Z',
   createdByApplicationId: '757f191a810c19729de860ae',
   createdByUserId: '757f191a810c19729de860ae',
   result: {
     outputOne: 'value',
     outputTwo: 'value',
     outputThree: 'value',
+  },
+};
+
+export const directExecutionResponseWithSnakeCasedCustomData = {
+  id: '757f191a810c19729de860ae',
+  status: 'new',
+  statusChangedTimestamp: '2023-07-03T14:14:50.698Z',
+  functionName: 'test',
+  data: {
+    input_one: 'value',
+    input_two: 'value',
+    input_three: 'value',
+  },
+  startTimestamp: '2023-07-03T14:14:50.698Z',
+  tags: ['nightly', 'test'],
+  priority: 1,
+  creationTimestamp: '2023-07-03T14:14:50.698Z',
+  updateTimestamp: '2023-07-03T14:14:50.698Z',
+  createdByApplicationId: '757f191a810c19729de860ae',
+  createdByUserId: '757f191a810c19729de860ae',
+  result: {
+    output_one: 'value',
+    output_two: 'value',
+    output_three: 'value',
   },
 };