+++ tests/unit/client/utils.test.ts
@@ -9,7 +9,7 @@ import {
 
 describe('recursiveMap function', () => {
   it('should recursively map with simple object', () => {
-    const result = recursiveMap(value => `-> ${value}`)({
+    const result = recursiveMap(value => `-> ${value}`, {
       test: 'value',
       groupIds: ['testGroupIds'],
     });
@@ -18,43 +18,13 @@ describe('recursiveMap function', () => {
   });
 
   it('should recursively map with object with arrays', () => {
-    const result = recursiveMap((value, key) => {
-      if (key.includes('stamp')) {
-        return new Date(value);
-      }
-      return value;
-    })({
-      id: '5bfbfc3146e0fb321rsa4b28',
-      first_name: 'John',
-      last_name: 'Doe',
-      email: 'john.doe@example.com',
-      activation: true,
-      phone_number: 9021324354,
-      profile_image: 'string',
-      language: 'NL',
-      time_zone: 'Europe/Brussels',
-      subObject: { expiry_timestamp: 1543240753289 },
-      patient_enlistments: [
-        {
-          group_id: '5bfbfc3146e0fb321rsa4b28',
-          expiry_timestamp: 1543240753289,
-          expired: true,
-          creation_timestamp: 1543240753200,
-        },
-      ],
-    });
-    expect(result.patient_enlistments[0].expiry_timestamp).toStrictEqual(
-      new Date('2018-11-26T13:59:13.289Z')
-    );
-  });
-
-  it('should recursively map an array containing object with arrays', () => {
-    const result = recursiveMap((value, key) => {
-      if (key.includes('stamp')) {
-        return new Date(value).toISOString();
-      }
-      return value;
-    })([
+    const result = recursiveMap(
+      (value, key) => {
+        if (key.includes('stamp')) {
+          return new Date(value);
+        }
+        return value;
+      },
       {
         id: '5bfbfc3146e0fb321rsa4b28',
         first_name: 'John',
@@ -74,20 +44,85 @@ describe('recursiveMap function', () => {
             creation_timestamp: 1543240753200,
           },
         ],
+      }
+    );
+    expect(result.patient_enlistments[0].expiry_timestamp).toStrictEqual(
+      new Date('2018-11-26T13:59:13.289Z')
+    );
+  });
+
+  it('should recursively map an array containing object with arrays', () => {
+    const result = recursiveMap(
+      (value, key) => {
+        if (key.includes('stamp')) {
+          return new Date(value).toISOString();
+        }
+        return value;
       },
-    ]);
+      [
+        {
+          id: '5bfbfc3146e0fb321rsa4b28',
+          first_name: 'John',
+          last_name: 'Doe',
+          email: 'john.doe@example.com',
+          activation: true,
+          phone_number: 9021324354,
+          profile_image: 'string',
+          language: 'NL',
+          time_zone: 'Europe/Brussels',
+          subObject: { expiry_timestamp: 1543240753289 },
+          patient_enlistments: [
+            {
+              group_id: '5bfbfc3146e0fb321rsa4b28',
+              expiry_timestamp: 1543240753289,
+              expired: true,
+              creation_timestamp: 1543240753200,
+            },
+          ],
+        },
+      ]
+    );
     expect(Array.isArray(result)).toEqual(true);
     expect(result[0].patient_enlistments[0].expiry_timestamp).toBe(
       '2018-11-26T13:59:13.289Z'
     );
   });
+
+  it('should skip the ignore keys', () => {
+    const result = recursiveMap(
+      (value, key) => {
+        if (key.includes('stamp')) {
+          return new Date(value).toISOString();
+        }
+        return value;
+      },
+      {
+        id: '5bfbfc3146e0fb321rsa4b28',
+        expiry_timestamp: 1543240753289,
+        custom_fields: {
+          group_id: '5bfbfc3146e0fb321rsa4b28',
+          expiry_timestamp: 1543240753289,
+          expired: true,
+          creation_timestamp: 1543240753200,
+        },
+      },
+      ['custom_fields']
+    );
+    expect(result.expiry_timestamp).toBe('2018-11-26T13:59:13.289Z');
+    expect(result.custom_fields.expiry_timestamp).toBe(1543240753289);
+    expect(result.custom_fields.creation_timestamp).toBe(1543240753200);
+  });
 });
 
 describe('recursiveRenameKeys function', () => {
   it('should map keys for simple object', () => {
-    const result = recursiveRenameKeys(value => `test_${value}`, {
-      test: 'value',
-    });
+    const result = recursiveRenameKeys(
+      value => `test_${value}`,
+      {
+        test: 'value',
+      },
+      []
+    );
 
     expect(Object.keys(result)).toStrictEqual(['test_test']);
   });
@@ -121,7 +156,8 @@ describe('recursiveRenameKeys function', () => {
             creation_timestamp: 1543240753200,
           },
         ],
-      }
+      },
+      []
     );
     expect(result.patientEnlistments[0].expiry_ts).toBe(1543240753289);
   });
@@ -157,7 +193,8 @@ describe('recursiveRenameKeys function', () => {
             },
           ],
         },
-      ]
+      ],
+      []
     );
     expect(result[0].patientEnlistments[0].expiry_ts).toBe(1543240753289);
   });
@@ -172,9 +209,12 @@ describe('camelize function', () => {
 
 describe('camelizeKeys function', () => {
   it('should camelize keys of an object', () => {
-    const result = camelizeKeys({
-      easy_string_to_test: 'test',
-    });
+    const result = camelizeKeys(
+      {
+        easy_string_to_test: 'test',
+      },
+      []
+    );
     expect(result.easyStringToTest).toBeDefined();
   });
 });
@@ -188,9 +228,12 @@ describe('decamelize function', () => {
 
 describe('decamelizeKeys function', () => {
   it('should decamelize keys of an object', () => {
-    const result = decamelizeKeys({
-      easyStringTo_test: 'test',
-    });
+    const result = decamelizeKeys(
+      {
+        easyStringTo_test: 'test',
+      },
+      []
+    );
     expect(result.easy_string_to_test).toBeDefined();
   });
 });