+++ tests/unit/configurations/generalService.test.ts
@@ -7,6 +7,7 @@ import {
   rqlBuilder,
 } from '../../../src/index';
 import {
+  customGeneralConfigResponse,
   generalConfig,
   generalConfigResponse,
 } from '../../__helpers__/configuration';
@@ -49,6 +50,22 @@ describe('Configuration: General Service', () => {
     expect(res.groupConfiguration).toBeDefined();
     expect(res.staffConfiguration).toBeDefined();
     expect(res.patientConfiguration).toBeDefined();
+  });
+
+  it('Should not transform custom data in the get response', async () => {
+    // const customResponseKeys = ['data', 'userConfiguration', 'groupConfiguration', 'staffConfiguration', 'patientConfiguration'];
+    nock(`${host}${CONFIGURATION_BASE}`)
+      .get('/general')
+      .reply(200, customGeneralConfigResponse);
+
+    const response = await sdk.configurations.general.get();
+    expect(response).toStrictEqual({
+      ...customGeneralConfigResponse,
+      creationTimestamp: new Date(
+        customGeneralConfigResponse.creationTimestamp
+      ),
+      updateTimestamp: new Date(customGeneralConfigResponse.updateTimestamp),
+    });
   });
 
   it('should update the general configuration', async () => {