+++ tests/unit/tasks/schedules/post.ts
@@ -17,18 +17,18 @@ describe('Tasks - Schedules - POST', () => {
   it('Creates a Schedule', async () => {
     nock(`${host}${TASKS_BASE}`)
       .post(`/schedules`)
-      .reply(200, { data: scheduleCreation });
+      .reply(200, scheduleCreation);
 
-    const { data } = await exh.tasks.schedules.create(scheduleCreation);
+    const data = await exh.tasks.schedules.create(scheduleCreation);
     expect(data).toMatchObject(scheduleCreation);
   });
 
   it('Creates a Schedule using a user defined data type', async () => {
     nock(`${host}${TASKS_BASE}`)
       .post(`/schedules`)
-      .reply(200, { data: scheduleCreation });
+      .reply(200, scheduleCreation);
 
-    const { data } = await exh.tasks.schedules.create<ScheduleDataType>(
+    const data = await exh.tasks.schedules.create<ScheduleDataType>(
       scheduleCreation
     );
     expect(data).toMatchObject(scheduleCreation);